LayoutTests/fast/events/transformaction/create-transformaction-event-expected.txt

This tests whether the DOM can create TransformActionEvents.
PASS event is non-null.
PASS successfullyParsed is true

TEST COMPLETE
